audio: NTS with MS12, ms12 insert zero when input ddp-atmos [1/1]
PD#SWPL-3936

Problem:
1.None-tunnel DDP-Atmos mode,
Audio Format is DDP-Atmos, and it's an sine-wav.
But the output is interrupt with silence frame.
2.Tunnel DDP-Atmos mode,
the input is interrupted with silence frame
in Netflix, logcat -s MediaDecoder2AudioPassthru
got the log:filldata, fillCompressedAudioData...

Solution:
1.get the frames by PCM nbytes/nbytes_per_sample,
the output nbytes of PCM from Dolby-MS12,
its api is dolby_ms12_get_n_bytes_pcmout_of_udc();
2.accurately calc the audio input time,
the method is to split one ddp frame to several pieces(a/b/c/d),
summary the duration of these pieces,
the total time is equaled with one whole dolby frame.

Verify:
U212 with NTS on GTVS.

Change-Id: I4c78b330f3ac425d2c6a27d05d0c6073b64e33d2
Signed-off-by: Xiushan Lu <xiushan.lu@amlogic.com>