audio: add constrain to open loopback device [1/1]
PD#SWPL-43327

Problem:
If USB_KARAOKE macro is defined but the karaoke is not on,
then the output port is alway trying to open the loopback device
for AEC. This is a waste of CPU loading which may bringin audio
underrun randomly.

Solution:
Add constrains that only the karaoke is on then try to open the
loopback device.

Verify:
U212.

Signed-off-by: Shuai Li <shuai.li@amlogic.com>
Change-Id: I0056526acbc1f8a1e7b822e7a78a203d1222cc49