audio: Fix EARC no sound issue when spdif out dd format [1/1]
PD#SWPL-63994

Problem:
When connect DD only AVR, there is no sound

Solution:
When AVR is not connected, it will output
dd format on spdif port.
But when connected DD only AVR, the spdif
output will not be closed. And the EARC port
can't get the correct info, so there is no sound.
Now when arc is connected, we need reset all
the spdif output.

Verify:
T3

Signed-off-by: yujie.wu <yujie.wu@amlogic.com>
Change-Id: Ied535f287bee43498e758fa31a868b05bcb9e1dc