audio: add tuner framework audio patch vol control  [1/1]
PD#SWPL-59737

Problem:
the volume can not controlled when tuner framework dtv

Solution:
for tuner framework pass through mode, the Audio HAL DTV
patch is triggered by android AudioTrack interface addtional
info, which is similar as direct/offload bistream/pcm MM
playback, so the volume will be also controlled through audiotrack.
The audio HAL outs_set_volume is invoked in this case.
In this case ,the real Audio track output stream is not the same
as the stream of the out_set_volume.So it can not applied.
As we already have dtv volume parameter maintained by dtv patch.
we can use this to apply that. also for MS12 enabled use case,
It is the same as the MM playback, use MS12 main volume control
interface to control that.
This volume control should not be stored when the AudioTrack is
destoryed as the EOL will go with current stream.
after the audio patch released, the volume will be restored to 1.0

Verify:
ohm

Signed-off-by: Jian Xu <jian.xu@amlogic.com>
Change-Id: Ia21b97be6e283b897c2ab7bb9944b08d6ea07d1d