audio: Fix DTV AV sync issue [1/1]
PD#SWPL-66744

Problem:
When hotplug HDMI, there is AV sync issue

Solution:
When hot plug HDMI, ms12 will be init, and
the audio data offset and pts is not match
in ms12. Now when prepare the ms12, we need
consider the input data offset, and set is
as a base.

Verify:
ohm

Signed-off-by: yujie.wu <yujie.wu@amlogic.com>
Change-Id: I2376f2a7f2f05613986b4920bc57c091a969e9d0