audio: MS12V2+DAP+LOGO the atmos does not popup when set virt ON. [1/1]
PD#SWPL-74705

Problem:
When the stream is playing, set the Speaker Virtualizer to ON.
Found that the Atmos popup did not pop up.

Solution:
Only for Dolby Atmos content, when the Speaker virtualizer is changing,
such as ON/AUTO -> OFF or OFF -> ON/AUTO, will update the infor again.

Verify:
need test with more test.

Change-Id: Iba492917f434aefac1c331b86411dafe0e7a4d82
Signed-off-by: Xiushan Lu <xiushan.lu@amlogic.com>