audio: MS12V1+Mixer high speakers have no sounds when atmos play [1/2]
PD#SWPL-76335

Problem:
{Audio}[R][MS12v1][Block][DTV]high speakers have no sounds
when atmos playback.3/3.

Solution:
when goto scan the dtv program, the ms12 works in the continous mode.
then adev->system_app_mixing_status is ON specially for ott mode. but
the height speaker is wrong for DDP_JOC content in Dolby ATMOS AVR,
if the system mixing status is on, the height speaker is mute.
if the system mixing status is off, the height speaker has sound.

Verify:
wait for further test.

Change-Id: Ia6bd6761b9a3201dc3254c2d6695be1e0037c733
Signed-off-by: Xiushan Lu <xiushan.lu@amlogic.com>