audio: No audio when DUT audio set to DD with optical [1/1]
PD#SWPL-80780

Problem:
No audio when DUT audio set to DD with optical.

Solution:
In the TV case, when pressed the KEY_MUTE and connected the soundbar,
choosing the DD/AUTO/PASSTHROUGH, we should not mute the spdif data,
so when sink[adev->active_outport] == 0.0, we should not mute the
spdif->b_mute.

Verify:
t3

Signed-off-by: binbin dai <binbin.dai@amlogic.com>
Change-Id: If519145939eda0101ca56fbba9e386daee61ea29