audio: Fix DTS HDMI-In noise when framesize is 2013 bytes. [1/1]
PD#SWPL-85931
PD#TV-57711

Problem:
Decoder check DSYNC fails and discards
this frame makes noise.

Solution:
If framesize is 2013(Pd 0x3EE8 >>3 ) and
ASPF(a.k.a Audio Sync Word Insertion Flag) is 1.
The last byte will be used as a parameter for DSYNC
verification and needs to be given to the decoder together.
Due to the endian problem, 2014 bytes need to be passed in,
and the one extra byte, the decoder will discard it.

Verify:
T3 R

Change-Id: I9f4e2bbf0a35f486002bf69c95b76b57d4958a51
Signed-off-by: sijia.lin <sijia.lin@amlogic.com>