audio: There is no sound output in ms12 case [1/1]
PD#SWPL-82390

Problem:
For HDMI IN case,there is no sound output in
ms12 process.

Solution:
In this case,the configuration(missing the "-oms") of ms12 is wrong.
This issue is fixed after changing the configuration, added the
stereo pcm out(-oms).

Verify:
Test on board T5W. It works well now.

Signed-off-by: wei huang <wei.huang@amlogic.com>
Change-Id: I107066aebf24bbb8ebbf385c9657369b88caa947