audio: Fix a bug where looping DTS stream caused video to stuck. [1/1]
PD#SWPL-84671

Problem:
Play the DTS stream in a loop, the audio PTS is not updated
and the player does av sync, causing the video to stuck.

Solution:
When playing DTS stream on non-ms12 pipeline, if ms12 is
not initialized, continue_mode needs to be turned off manually.

Verify:
Ohm, Android S/R

Change-Id: I3eb7bcb91577784874accacfc395b28336ebdf9b
Signed-off-by: sijia.lin <sijia.lin@amlogic.com>