audio: There is no sound output for eARC after switch output format [1/1]
PD#SWPL-89848

Problem:
There is no sound output for eARC after switching between
"Passthrough" and "AUTO".Actually,it is two issue for this
jira.The root cause is different. But the observation of
issue is the same.One is for box device.There is over run
happen for pcm device of parse.So it causes parse fail.This
issue happen.Another issue is for DVD device.For DVD device,
the packet change message is different with box device.For box,
packet changes from source value to the target value directly.
But for DVD device,the value changes with several phase. Finally,
it shifts to the target value.Now it bases on this message to reopen
pcm device.It gives wrong channel config during DD+ playback
when use DVD.So it causes data transfer by using 8 channel.So
it parses fail.

Solution:
Change period count to enlarge DMA buffer size to avoid over run.
Add one more case to judge whether it needs to recover config
of pcm device.

Verify:
Test on board T3. It works well now.

Signed-off-by: wei huang <wei.huang@amlogic.com>
Change-Id: If8b7a021ea81abf60c2f4ec4c07b16a7826c6a88