audio: stream Loop playback, sound output is not smooth [1/1]
PD#SWPL-102301

Problem:
stream Loop playback, sound output is not smooth.

Solution:
1.If the available main size is less than the main frame
size(1636 or 768) when find the main header and have got
the valid frame size, there will memset the data. However,
we do not memset this data and need to save the main head and
read size which is ensure to confirm the integriry data.
2. If ms12 main process can not find find the iec61937 format
header, we need drop this error frame data, If not, there will
continuous sleep until find the iec61937 format header which
will make 5s and effect the audio output when stream loop back..

Verify:
T5D

Change-Id: I0a6f8b54987c79ea6a68c4855664790845359ef1
Signed-off-by: binbin dai <binbin.dai@amlogic.com>