audio: fixed pvr ad stuck issue [1/1]
PD#SWPL-102215

Problem:
open AD switch,play PVR video,AD audio track
is not smooth,Frequency:2/2,Comparison:none

Solution:
when pvr playback , the cache of main data
may be more than the cache of ad data,that
lead to can not align the data of ad and mian.
so need to wait the ad data but do not effect
the main data output.

Verify:
ohm

Change-Id: If26e2b92ce188e3046199a23e34fbe8f0f448b56
Signed-off-by: lianlian zhu <lianlian.zhu@amlogic.com>