audio_hal: fix unable to control bt volume [1/1]
PD#TV-71955

Problem:
The sound of xiaomi soundbar does not change when the volume is adjusted

Solution:
1. Some Bluetooth speakers do not support avrcp and will not set the
avrcp status to audio hal after connected.
2. When platforms first connect to avrcp bt, disconnect
and then connect to do not support avrcp and do not emit
avrcp support status, audio hal believes that the AvRCP-enabled speakers
are causing the volume to become uncontrollable.

Verify:
smith + T

Change-Id: Ice816b507be406bd0e22dd9f1fedb3c1a205ef36
Signed-off-by: hongchao.yin <hongchao.yin@amlogic.com>