audio: Fix DTS HDMI-In noise when framesize is odd number. [1/1]
PD#SWPL-111258
PD#TV-70386

Problem:
Decoder check DSYNC fails and discards this frame makes noise.

Solution:
If framesize is a odd number, (such as 2013(Pd 0x3EE8 >>3 ))
and ASPF(a.k.a Audio Sync Word Insertion Flag) is 1.
The last byte will be used as a parameter for DSYNC verification
and needs to be given to the decoder together.
Due to the endian problem, one more bytes need to be passed in,
and the one extra byte, the decoder will discard it.

Verify:
T5D, Shine

Change-Id: I728a25a96c62f06e2f86b88bcf8206eca1e9b880
Signed-off-by: sijia.lin <sijia.lin@amlogic.com>