audio: There is noise when resume playback for DDP stream, 3/5 [1/1]
PD#SWPL-114715

Problem:
It uses software parse for data format in HDMI case.It is slow
to parse the data type when use this method.It does sample rate
convert for the input raw data and plays in pcm mode.So it causes
this noise issue.

Solution:
Keep previous data type in pause status.

Verify:
Test on board T7C. It works well now.

Change-Id: I45e939126fe1b554e697f85d92b7824a52253cbc
Signed-off-by: wei huang <wei.huang@amlogic.com>