audio: fixed dtv and netflix switch blocked issue [1/1]
PD#SWPL-116244

Problem:
When switch cbs and netflix, sometimes the dtv is
blocked.
The normal case is that  cbs mediasync and netflix
mediasync  is different. but cbs and netflix apk
get hwsync id from common audio hal api adev_get_para.
the abnormal case is that cbs get mediasync and start running ,
then netflix mediasync get the mediasync id, it is the same with
cbs mediasync id, and then release the netflix mediasync,so the
cbs mediasync is paused

Solution:
we add a patch to check whether the sync id is matched.
If it is  matched, we do not need to release netflix mediasync .
If app can solve the sequence issue, then this patch is not needed.

Verify:
ohm

Change-Id: I7fd0119cbfa69f57f808448706bcec85b3b296b3
Signed-off-by: lianlian zhu <lianlian.zhu@amlogic.com>