audio: The local video sounds abnormal after seeking [1/1]
PD#SWPL-118414

Problem:
For speaker output, the data needs to be processed by VX effect.
There is buffer in VX library. So it saves data which be processed
of previous stream. When start one new stream, the data of previous
stream will be played in current stream. So it causes this issue.
And it also happens when do pause operation if playing DTS stream
in HDMI case.

Solution:
When close or pause one stream, input silence data to VX library
to clear buffer if the output is speaker.

Verify:
Test on board T3. It works well now.

Change-Id: Ib894634ed4696daeadfa06c620f3e01f3afa941e
Signed-off-by: wei huang <wei.huang@amlogic.com>