audio: edid updating failed [1/1]
PD#SWPL-128158

Problem:
In auto mode,the ECE interaction sequence of MAT, DDP and AC3
audio formats is C, A and 2.The last time is the AC3(2) format,
which updates EDID to the default EDID, so MAT_PCM and ATMOS
are not supported.

Solution:
Update MAT (MAT_PCM) EDID information and DDP (ATMOS) EDID information
respectively when the CEC interaction information of MAT and DDP is
obtained; When the CEC interaction information of AC3 is obtained, the
EDID information is updated to the existing EDID information.

Verify:
T5W

Change-Id: Id2b66726ece26ed4f29e3a1dd305004e390bfd37
Signed-off-by: shaohui.sheng <shaohui.sheng@amlogic.com>
(cherry picked from commit c2c221fd1b04383626251649dcfc912774f46cb3)