audio: Switch to the ad track, no sound output [1/1]
PD#SWPL-148690

Problem:
Switch to the ad track, no sound output.

Solution:
The normal size of metadata is 256 and will find
the valid metadata header. However, if the size of
metadata is 512 which is obtain two valid packet,
it will lost a valid packet. Therefore, when we find
the valid metadata, it need break the loop in time.

Verify:
U OHM

Change-Id: Iefafdcfc3e3363c5858e9fad5389ab427de6dce8
Signed-off-by: binbin dai <binbin.dai@amlogic.com>