audio: HDMI playback not smooth after ARC plugging in and out [1/1]
PD#SWPL-155064

Problem:
For stream with 44.1khz, it may insert 48Khz silence
frame when do avsync.Sometimes it will parse
this 48Khz silence frame first and then open ALSA
device with this sample rate. So it causes this issue.

Solution:
For ARC output, it is not suitable to insert silence
frame in avsync code logic.Remove this process for
ARC device.

Verify:
Test on board T3X. It works well now.

Change-Id: Icee206e12b7ebfe01634cefb941fb202266ae1e8
Signed-off-by: wei huang <wei.huang@amlogic.com>