audio: fix DPE effect Crash [1/1]
PD#SWPL-165747

Problem:
DPE algorithm can only process 256 frames of audio data
each time, less than 256 frames will be cached and the
next data to be processed into 256 frames to continue
processing, when switching, the size of the first data
in the incoming algorithm library and the buffer size
of the incoming algorithm library, and there are less
than 250 frames of data in the algorithm library. In
this way, when the data was spliced, the boundary was
crossed, resulting in a crash.

Solution:
Increases the buffer size in the incoming library by
an additional 256 frames of data to cache the remaining
data from the last processing of less than 256 frames.

Verify:
test on txhd2

Change-Id: I25f8a5166efea70864112602f982c9a15b3d4f63
Signed-off-by: shaohui.sheng <shaohui.sheng@amlogic.com>