audio: Optimize Game mode libsync again [1/1]
PD#SWPL-159180

Problem:
The audio latency is too large in game mode.

Solution:
In game mode, the ALSA output device is turned on again.
At the same time, the output DMA buffer is configured to
the minimum possible under the condition of normal playback,
that is, the latency generated by the output DMA buffer is
the minimum.

Verify:
test on T3X

Change-Id: Id67345ad9c36dfa0423a595db0b391dd2d681b95
Signed-off-by: shaohui.sheng <shaohui.sheng@amlogic.com>